refactor(Types.Code): to Types.CodeMirror
