+++ admin/public/styles/keystone/field-types.less
@@ -5,7 +5,7 @@
 
 
 
-// Code
+// CodeMirror
 // ------------------------------
 
 .CodeMirror-container {
