+++ admin/server/templates/includes/css/components.jade
@@ -1,2 +1,2 @@
-if list.fieldTypes.code
+if list.fieldTypes.codemirror
 	link(rel="stylesheet", href="#{adminPath}/js/lib/codemirror/codemirror.css")
