+++ admin/server/templates/includes/script/components.jade
@@ -3,7 +3,7 @@ script(src="#{adminPath}/js/lib/move/move-0.1.1.min.js")
 script(src="#{adminPath}/js/lib/jquery-placeholder-shim/jquery-placeholder-shim.js")
 if list.fieldTypes.wysiwyg
 	script(src="#{adminPath}/js/lib/tinymce/tinymce.min.js")
-if list.fieldTypes.code
+if list.fieldTypes.codemirror
 	script(src="#{adminPath}/js/lib/codemirror/codemirror-compressed.js")
 if cloudinary
 	script(src='#{adminPath}/js/lib/jqueryfileupload/vendor/jquery.ui.widget.js')
