+++ lib/fieldTypes.js
@@ -3,7 +3,7 @@ var fields = {
 	get Boolean () { return require('../fields/types/boolean/BooleanType'); },
 	get CloudinaryImage () { return require('../fields/types/cloudinaryimage/CloudinaryImageType'); },
 	get CloudinaryImages () { return require('../fields/types/cloudinaryimages/CloudinaryImagesType'); },
-	get Code () { return require('../fields/types/code/CodeType'); },
+	get CodeMirror () { return require('../fields/types/codemirror/CodeMirrorType'); },
 	get Color () { return require('../fields/types/color/ColorType'); },
 	get Date () { return require('../fields/types/date/DateType'); },
 	get DateArray () { return require('../fields/types/datearray/DateArrayType'); },
